services/paypal-payments/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic-settings
aio-pika
pydantic[email]
aiohttp
watchfiles
databases[postgresql]
asyncpg
sqlalchemy